Put screen buffers into place
We'll write out graphic chanegs to a front buffer in regular RAM.

This will be compared with the back-buffer (the previous version of the front buffer), and any changes will be queued up to be made during vblank.

Right now, we have a simple routine to copy the whole buffer to VRAM, rather than do a partial update.

Made a "gb" folder for GameBoy-specific stuff.

Put some teset graphics on the screen.