Register shuffling in `PointToCard`
I'm in the process of making the cursor point to different places based on the type of pile and card, but to avoid some gnarly stack use I'm going to have to move the pile card-count to the pile struct, rather than where it currently is, in the pool.

This commit gets us to a stable point before I make that change.