Begin work on card-sprites
We want to build a card stack from sprites, but sprites are 8x16 and sprite tiles can come only from either the sprite table or the *first half* of the background tiles.

To this end, we fill the sprite tiles with 8x16 pieces of the cards and the lower-half of the bg tiles will be used for the card back.

Whilst this duplicates the card tiles, it does away with the need to manually draw a card stack into the sprite tiles, something that is very difficult to do on the DMG without it's tile DMA.

I do however need to switch to using the GB's sprite OAM DMA so that I can just setup the card-stack sprite in RAM without having to wait for VRAM access.