WELL THAT WASN'T WORTH THE EFFORT
Not entirely true, but trying to order the tiles in the background tile set to be shared with sprites caused a lot of headaches with tile ordering and indices.

That said, palettes and transparency are correctly sorted now and it does highlight the fact that the cards might need outlines when used as sprites.