Build sprites from top down
When creating a sprite version of a card, start building sprites from no.40 downward; this will be needed for the overlapping cards (as all sprites are 8x16).

"cursor.wla" was removed by moving its contents into "sprites.wla".

We still need to implement moving the cursor with cards in hand...