Switch up the way we define Pile types
This is to make drawing piles easier and to help with validating cards in piles, according to their properties.

I want to make room in the type byte to hold a value for pile validation, e.g. if the cards stack in alternating colour etc. I could use another bitmask value for these properties, or just have a lookup table given x number of possible pile types; we shall see.