Validate lowest card of each suit in Aces Up
We walk the piles and validate the cards, marking them as "valid" or "invalid" which will be used to 'grey-out' invalid cards and control pointer interaction (can't pickup invalid cards).

Only Aces Up validation is supported right now, but this is easily expanded.

The bundled version of WLA-DX was updated to fix a label bug.