Swap data-orientation
Navigating up and down the fields in the piles was not scaling well and causing a lot of register contention. I've rewritten the game to use column-orientation instead, where each field is its own array; that is instead of each pile being self-contained (and thus having to walk up and down the fields to read them), each pile is now a row in a table, stored together as a column. We can freely jump between columns without having walk fields one by one.

In practice, this looks like this: field --> `[HL]` <-- pile index. To change which field being read, just change `H`. To change the pile being read, change `L`.

There are number of benefits to this:
1. we no longer have to `inc`/`dec` up and down the pile struct to read its data -- we can jump directly to any field just by changing `H`

* We can iterate over piles easily by index only. This also means that we can use the pile index as the counter and save on another register

Doing this change has *massively* simplified the code base and generally reduced register contention everywhere. Some cleanup still needs to be done to re-optimise code under this new scheme, but this commit represents the minimum required to get the game working as it was before.

A large number of bugs were fixed during the swap, including the deck not being populated with cards correctly.