Remove double-spacing in BG tilemap
This was done first to share the tiles with sprites, but since I went with a unique set of card tiles for the sprites (black outline), the double-spacing is no longer needed.

This also opens up space for 'red' versions of the ordinals that I had forgotten thus far. The tiles are included, but not wired in yet.