Wire up black|red ordinals
The tiles for the ordinals are now flanked by "red" versions. This necessitated moving the "colour" bit (black or red suit) to the right so that the suit occupies the 5th bit of the rank.

On the GBC the ordinal colour can be done with palettes, but that's a detail for another day.