Begin rework of validation
I'm not a fan of greying out of cards that cannot be played as it's visually confusing in some games, but also there's many differences between which cards are free to be picked up vs. what is a valid drop target, and where the cursor should/shouldn't move to in piles.

Some `PILE_*` constants were changed to `TYPE_*`.