Begin adding gameplay, i.e. automatically moving a card once tapped.
The game can now respond to tapping on a card and act according to the game rules. In most cases this will be to move a card from one pile to another.

Added a cursor mode to move individual cards around the screen.

The building of card sprites was moved to `sprites.wla`, some refinement needed. We need to use different position of the sprites when using the cursor hand or not.