Change cursor to a card
For moving a card automatically from one pile to another, we need to build a sprite version of the card, but not held by the cursor as with pick/drop-pile.

This commit goes as far as constructing a card sprite in the same place as the card in a pile, but doesn't remove the card from the pile yet.

`rst_lastCardAddr` & `rst_cardAddr` now return the card index.

`getPosCard` added to return the tilemap row/col of a specific card in a pile, accounting for horizontal or vertical spread.

A flag has been added (`update_sprites`) to indicate when the sprite attribute table has changed; this is done so that the hardware sprites do not get updated mid-way through a change in the working attribute table.

A number of routines in cursor.wla were moved to tableau.wla because they pertain to pile manipulation, which is specific to gameplay, and wouldn't be relevant when using the cursor in the menu, for example. More routines need to be moved, but I didn't want to make this commit too big.