Move a card (visually) betwen piles
The vblank/screen RSTs were moved out to regular routines since these are not the most common calls and I needed to free up another RST.

There's a problem with the cursor sprite not being visible when moving the card sprite, so I need to fix the handling of the different cursor modes. We also need to disable input during the move, or in other situations. A local spin-loop may do during the move animation.