Improvements to Menu code
A return value is provided by the menu to indicate which menu item was activated (or if the menu was cancelled)

Actually doing the chosen menu action yet to come!

The input loop was also set up so that changing the game from the menu would return to the input loop naturally, without the need to force a hard jump and reset of the stack.

Added a backup of the seed when a game is started. This will be used to be able to replay the current game with the same cards (RNG).